split  skin shouldIgnore processors normalizeUrl into own files
